Supplemental telemetry definition support for C# (#47)
This change sets up the C# telemetry generator to work with supplemental telemetry definitions. Supplemental telemetry definitions are product specific definitions that (typically) live within a products repo. For example, the VS Toolkit may have some VS Toolkit specific telemetry metrics, which don't make sense to add to the common definitions contained in this repo.

The intention is that (in upcoming changes) generated code for the common telemetry definitions is packaged (along with supporting datatypes), and that package is consumed by the toolkit. Separate code is generated from supplemental telemetry definitions, which is directly added to the toolkit, and has a dependency on the package containing the common definitions and supporting datatypes. There are tests in AwsToolkit.Telemetry.Events.Tests to illustrate this.

Also, I made changes to get the C# CodeBuild check working again (pending changing the CodeBuild image from linux to windows, now that we're using .NET Framework in addition to .NET Core).