Document awsAccount and awsRegion as metrics properties; add them to C# generator (#106)
This change documents `awsAccount` and `awsRegion` as properties available to all metrics.

This change also updates the C# telemetry generator so that these properties are available to the VS Toolkit. I also added some constant values ("n/a", "not-set", "invalid") to have some consistency with the JetBrains implementation.

Updates to `telemetry/csharp/AwsToolkit.Telemetry.Events.Tests/Generated/GeneratedCode.cs` are autogenerated and used in tests. That file is illustrative of what the generated telemetry code looks like, but also accounts for changes made to the telemetry definitions since the last time someone worked in the C# generator code.