Update C# telemetry generator to allow client code to adjust if a metric is passive (#108)
The VS Toolkit hasn't had a need to designate certain metrics as either passive or active. #107 identifies a scenario (`codecommit_setCredentials`) where we have both active and passive activities raising this event.

This change updates the C# generator to allow client code to adjust the `Passive` state of a metric. To accomplish this, a field "Passive" has been added to the base class for all metric classes (`BaseTelemetryEvent`). This class is initialized within every metric class constructor.

Updates to `telemetry/csharp/AwsToolkit.Telemetry.Events.Tests/Generated/GeneratedCode.cs` are autogenerated and used in tests. That file is illustrative of what the generated telemetry code looks like, but also accounts for changes made to the telemetry definitions since the last time someone worked in the C# generator code.