Add credentialSourceId to set credentials metric (#110)
This change adds an optional field to the "set credentials" metric, which allows us to determine where the credentials came from. This will give us a better indication of usage patterns across different credentials sources.

For example, in the VS Toolkit, the credentials profile can currently from from the encrypted SDK Credentials store, or the Shared Credentials files.

`windowsCredentialStore` has been removed from `credentialType` because it is a container for a credentials definition rather than a type.