Add optional handler to adjust metadata prior to logging metrics (#296)
This change updates the generated metrics code to provide an optional callback that allows the caller to transform the metric properties prior to sending them to the server. This could be used for example to chop a long value into smaller segments.

Updates to `telemetry/csharp/AwsToolkit.Telemetry.Events.Tests/Generated/GeneratedCode.cs` are autogenerated and used in tests. That file is illustrative of what the generated telemetry code looks like, but also accounts for changes made to the telemetry definitions since the last time someone worked in the C# generator code.

This change was co-authored with @floralph