Update C# Telemetry Generator to use .NET 6 instead of .NET Core 3.1 (#398)
This change updates the C# Telemetry Generator to target .NET 6 instead of .NET Core 3.1. .NET Core 3.1 reaches end of life this month.

This change also updates the GitHub action to address a Node 12 deprecation warning.

Updates to `telemetry/csharp/AwsToolkit.Telemetry.Events.Tests/Generated/GeneratedCode.cs` are autogenerated and used in tests. That file is illustrative of what the generated telemetry code looks like, but also accounts for changes made to the telemetry definitions since the last time someone worked in the C# generator code. This file does not need to be reviewed. I produced this file in a commit prior to making the rest of the changes, so that I could see that the code produced with .NET 6 is identical.