chore: Implement uri caching functionality
This commit sets up a caching mechanism
for downloading URIs. This is necessary since
we need to provide a schema file on every request
to the YAML language server and it will significantly
slow things down if we need to retrieve the schema
over the network each time.

The caching works by storing the content locally
in some directory and afer a certain timeout period
it will attempt to download the uri content again
if necessary.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>