chore: refactor for client + server dir
We want to separate the source directory
in to server and client directories.

The server will hold all the parts of
the language server.

The client will hold all the parts of
specific LSP clients such as vscode
or visual studio. This will allow for
easier testing of the language server
as a whole on the supported ide.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>