refactor: simplify YAMLLanguageService
Problem

Due to the design of the YAML Language Service
api, it adds complexity to how we have to use it.
Specifically, when we want to use a feature
(completion, hover, ...) we cannot pass in the
schema that we want to apply to it.

Instead, we must pre-define the schema mapping
and then on subsequent calls it will handle
the schema for us.

Since we want the contribution process to be as
simple as possible for users, we need to simplify
this or else they will be forced to write additional
boilerplate code.

Solution

Create a wrapper around the YAML Language Service
that handles the more complex operations that we
can hide from the user. This will make it more
simple to use the YAML Language service.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>