refactor: Create BaseLanguageService
Before this commit we implemented the LanguageService
interface.

The issue with this is that as we add new methods to the
LanguageService, some implementations will not support
all of them.

Instead of the implementation needing to create a method
for something they don't support, this commit creates
a new abstract class which can be extended. This abstract
class throws for all methods and relies on the extender
to implement what they are able to provide.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>