Convert the lsp "root" to a monorepo root, seed it with an empty core package (#495)
This is the first in a series of changes that remakes the `lsp` folder into a monorepo that we can perform experiments on.

I've left all of the existing code in place for this change, but we don't compile it. Upcoming changes will move or delete the existing code.

The monorepo has been created based using guidance from https://earthly.dev/blog/setup-typescript-monorepo/

An empty package `core/aws-lsp-core` has been set up as a supporting core library, so that CI has something to compile.

If/when we decide to start publishing packages, the repo will be updated with a formal namespace. For now it will be set to `@lsp-placeholder`