CloudFormation language server and binary (#502)
This is part of a series of changes that establishes an LSP monorepo that we can perform experiments on.

This change provides a second demonstration language server and binary that can be integrated into IDEs. This language server is based on the CloudFormation JSON Schema.

This server's implementation differs from the Buildspec implementation (from https://github.com/aws/aws-toolkit-common/pull/500), for the purpose of comparison. In this instance, the service is injected into the server. This looks like it gives us more design flexibility in reusing and composing the services to suit the needs of where they will be hosted/integrated.