Proof of concept language server that provides CodeWhisperer recommendations (#526)
This change adds a concept language server that provides code recommendations from CodeWhisperer. The recommendations are returned through code completions and inline completion.

The sample VS Code extension has been updated to support this language server when editing typescript files. This is accomplished by introducing an inline completion provider, which wraps the concept language server.

At this time, inline completion is not a part of the language server protocol. VS Code has an extensibility API for inline completion, and a proposal is part of the next protocol version 3.18 (https://github.com/microsoft/language-server-protocol/pull/1673). This proposal is modelled after the extensibility API.

A likely implementation of the protocol is proposed in https://github.com/microsoft/vscode-languageserver-node/pull/1190, and portions of this change leverage the shapes defined in that proposal. This way, if or when inline completion is a part of the spec, we could update code to reference `vscode-languageserver` related types with little effort.

This concept uses IAM credentials with CodeWhisperer as a way to get up and iterating quickly. In future work we will want to switch over to using bearer tokens, like the AWS Toolkit for VS Code does.

The CodeWhisperer service client was produced using the AWS Toolkit for VS Code's client generator (https://github.com/aws/aws-toolkit-vscode/blob/master/scripts/build/generateServiceClient.ts)

This language server is being submitted so that we can perform additional experimentation such as:
- verifying browser compatibility
- exploring the UX around surfacing code recommendations
- future bearer token support