telemetry(amazonq): source folder modification #731
amazonq_modifySourceFolder can happen on different steps of our process,
whenever we upload the repo to our backend (that happens at least once before
planning the execution, and once more for each code generation requested).

It's not triggered by the user on-demand; only when we think the repo is too
large and a sub-folder should be selected. We display the button and want to
capture the action of the user clicking it.

While the field could be added as a userHasModifiedSourceFolder boolean of our
amazonq_uploadRepo and amazonq_GenerateCode metrics, the idea was to have
a specific metric indicating that the button was clicked (and also to have an
easier way to visualize it instead of searching for other metrics that have this
property set).