fix(tests): noisy console.error() messages
Problem:
When parsing fails, it logs redundant errors and then throws an
exception. This is pointless and redundant because the exception can
carry the same info and anything showing the console output will also
show the error message.

And the console.error() calls add lots of noise to *successful* test
results (for tests that *expect* validation to fail):

    console.error
      Error while trying to parse the definitions file /not/a/real/path: {}

      55 |         return input as MetricDefinitionRoot
      56 |     } catch (errors) {
    > 57 |         console.error(`Error while trying to parse the definitions file ${fileName}: ${JSON.stringify(errors)}`)
         |                 ^
      58 |         throw new Error('Failed to parse')
      59 |     }
      60 | }

      at validateInput (src/parser.ts:57:17)
      at test/parser.test.ts:19:35
      at Object.<anonymous> (node_modules/expect/build/toThrowMatchers.js:74:11)
      at Object.throwingMatcher [as toThrowError] (node_modules/expect/build/index.js:320:21)
      at Object.<anonymous> (test/parser.test.ts:19:64)

    ...[many more]

Solution:
Remove the console.error() calls.