Apply CultureInvariant string conversion to more numeric metadata types (#781)
## Problem

Most metadata is emitted using code that is auto-generated. There are times where we want to emit metadata from a generic source (like a deserialized JSON payload). These fields are treated as generic objects, which means numeric values get converted to strings using the user's locale settings. We have overloaded methods for certain types, which handle culture invariant conversion, but these fields are routed to the `object` overload. We want these to be culture invariant.

## Solution

Add an overload that can be explicitly used, which routes the metadata to their type-specific handler.

This change also converts the one test class assertions to use FluentAssertions.