Update the C# Telemetry Generator from .NET 6 to .NET 8 (#799)
This change updates the .NET telemetry generator from .NET 6 to .NET 8. The NuGet packages it produces continue to target .NET 4.7.2 (which supports the Visual Studio Toolkit).

The motivation for this change is to migrate off .NET 6 before its LTS window closes. A while back, I updated the infrastructure so that .NET 8 is available in the automation and CI jobs.

This change was verified by:
- checking that the test suite passes
- comparing the resultant generated code before and after the change, and ensuring they match