feat(generation): allow `awsRegion` to be specified by callers in Metadata
Problem: `awsRegion` is a base metric and does not allow for callers to add it to their metadata when using `record()`. However, this field is often known best by the caller and not the telemetry guessing algorithm.

Solution: Update type generation to allow this base metric to be specified in metadata.