feat: Add `viewId` common definition type
This adds a new type called `viewId` which uniquely identifies
a part of the UI in an IDE window.

The `viewId` should be a comma delimited string where each item
appended is a more specific part of the UI within the previous
item.

The top level (first item) part of the string should be an identifier
for the specific IDE window, as it is possible to have multiple windows
of the same IDE opened at the same time. Then if there is a specific
view within that specific IDE window you would append its identifier.

Example:

- I have 2 VS Code windows with IDs: `abc-123` and `xyz-789`
- My Amazon Q chat view has an ID `amazonq-chat`
- Result `viewId`: `abc-123,amazonq-chat` and `xyz-789,amazonq-chat`

Now in my metrics I will be able to differentiate metrics by their viewId
since they both share the same clientId by design.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>