Add definitions for a feature's enabled state, and the opening and closing of a module (#835)
## Problem

We're revising the VS Toolkit's Getting Started experience to more
cleanly enable and disable features. We wish to understand the user's
journey through this experience, and identify when/why a feature gets
enabled or disabled.

## Solution

A generalized set of metrics that can be applied to any feature:
- `toolkit_featureState` - intended to be emitted at various points in a
workflow, to report whether or not a certain feature is enabled. Also
emits "why" this state is being emitted (via the "source" field)
- `toolkit_openModule` - intended to be emitted when a
feature/view/component of interest is opened/entered/displayed. "What"
is being opened is represented by the "module" field, and "Why" it is
being opened is represented in "source"
- `toolkit_closeModule` - intended to be emitted when a
feature/view/component of interest is closed/exited. "What" is being
opened is represented by the "module" field.


## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.