jetbrains: Generate OpenTelemetry-style metric builders (#899)
## Problem
At the moment, metrics in the JetBrains extensions are collected as
singular datapoints independent of each other. Modern observability
frameworks, as well as our long-term goals are to be able to correlate
these data flows over a single user flow. This requires the 'trace ID'
concept which is high-friction in the current model.

## Solution
Generate 'fluent builders' that wrap the OpenTelemetry SDK to allow us
to leverage the SDK to record trace/span IDs automatically, but still
expose a good amount of type safety based on the metric schema. This PR
references base classes proposed in
https://github.com/aws/aws-toolkit-jetbrains/pull/4982.

## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.