auth: add `sessionDuration` to `aws_loginWithBrowser`
We want to know about the previous sessionDuration, if there is one, the next time we log
in to the same SSO session due to something like expiration.

We will use this in CloudWatch to alarm on if we are seeing a scenario like
BuilderID emitting another `aws_loginWithBrowser` less than 90 days after the previous
`aws_loginWithBrowser`.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>