Add reasonDesc to C# telemetry system (#954)
`reasonDesc` is a "global" field that we treat as available on all
metric types. This change updates the C# telemetry so that events have a
`ReasonDescription` property, which is emitted to `reasonDesc` metrics.

When reviewing this change, you can largely ignore
telemetry/csharp/AwsToolkit.Telemetry.Events.Tests/Generated/GeneratedCode.cs
- This file is not updated each time telemetry definitions change. Since
I refreshed it, it contains a larger set of adjustments.

#764 is the equivalent change for typescript

## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.