vsc: add `source` to common fields
Problem:

The `source` value has become a common pattern in our telemetry which answers
"who" caused the metric event. For example if we had a function() to open
a menu, there could be many different entrypoints that could call it.

Solution:

`source` will answer "who" it was for that event. And now it will
be a part of all metrics as an optional field.

Note:

The other IDEs should consider following this pattern if not already.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>