telemetry:vsc: add `source` to common fields (#977)
## Problem:

The `source` value has become a common pattern in our telemetry which
answers "who" caused the metric event. For example if we had a
function() to open a menu, there could be many different entrypoints
that could call it.

Right now we have to add `source` to each metric that uses it, but this
usage has increased for a lot of metrics.

## Solution:

- `source` will be a part of all metrics as an optional field.
- This has only been implemented for VSC.

## TODO

The other IDEs should consider following this pattern if not already.
**If all IDEs implement this then we can remove the explicit
requirements of `source` in `commonDefinitions.json`**

<!---
    REMINDER:
    - Read CONTRIBUTING.md first.
    - Add test coverage for your changes.
    - Link to related issues/commits.
    - Testing: how did you test your changes?
    - Screenshots if applicable
-->

## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>