doc: update metric naming spec (#984)
## Problem

We do not have a clear defined format for metric names.

## Solution

Update the doc that defines how to name a metric.

- This will NOT retroactively apply to all existing metrics since this
will require significant work refactoring that we do not want to do at
the moment.

- The benefits of this new structure `{namespace}_{noun}{Verb}` are that
it will group related metrics in the definitions file since they are
alphabetically ordered. Also when getting IDE completions all relevant
options will be grouped together. This should help for
searchability/reusability.

<!---
    REMINDER:
    - Read CONTRIBUTING.md first.
    - Add test coverage for your changes.
    - Link to related issues/commits.
    - Testing: how did you test your changes?
    - Screenshots if applicable
-->

## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Signed-off-by: nkomonen-amazon <nkomonen@amazon.com>