telemetry: add additional context types to `amazonq_interactWithMessage` (#986)
## Problem
We want to understand if there is a change in user interaction with
messages that include the new falcon features. This will help us
evaluate the success of the features.

## Solution

- Add # of files, # of folders, # of prompts, and # of rules to the
amazonq_interactWithMessage metric

<!---
    REMINDER:
    - Read CONTRIBUTING.md first.
    - Add test coverage for your changes.
    - Link to related issues/commits.
    - Testing: how did you test your changes?
    - Screenshots if applicable
-->

## License

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.