telemetry(amazonq): list of request ids in addMessage #1024
## Problem
- The following are new science requirements:
- In the new agentic chat after a user types a message, it may trigger
multiple LLM calls, but we only store the final LLM requestID
(`cwsprChatMessageId`), which is the one that generates the response
shown to the user. Earlier LLM calls in the sequence, which handle
internal processing and aren't visible to users, are not currently
emitted.

- We currently emit the LLM interaction events, but we lose emission of
the associated LLM/Tool RequestID’s that led to the user interaction
event (ie we don’t know which LLM requests the user click stop button,
we don’t know which write tool ID the user chose to undo).

## Solution
- Emit all LLM RequestIDs in the order that these requests were made in
`amazonq_addMessage`
- As the list of RequestIDs may be long, will look into the max char
limit of a field in kibana
- Emit Tool/LLM requestIDs related to User Interaction Events in
`amazonq_interactWithAgenticChat`