Fix endpoint override for .NET SDK generator (#1088)
PR #1082 migrated the C# telemetry generator to AWS SDK V4. Due to
changes in contracts with this version, a Custom PartialConfig file was
added to meet the interface requirements for configuring endpoints with
the ClientConfig. This caused issues during runtime since our client
does not rely on endpoitns. As a workaround as per .NET SDK team's
guidance, since our client always sets the ServiceUrl on out client, a
fix along those lines has been made to short-circuit the underlying
logic and return endpoint based on the service url.